+++ demos/jt-demo-808-server-webflux-boot3/src/main/java/io/github/hylexus/jt/demos/jt808/JtDemo808ServerWebfluxBoot3.java
@@ -7,6 +7,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class JtDemo808ServerWebfluxBoot3 {
 
     public static void main(String[] args) {
+        // FIXME 如果你不了解 ResourceLeakDetector 是做什么的,请务必注释掉下面这行代码!!!
+        io.netty.util.ResourceLeakDetector.setLevel(io.netty.util.ResourceLeakDetector.Level.PARANOID);
         SpringApplication.run(JtDemo808ServerWebfluxBoot3.class, args);
     }